// src/app/model/product.model.ts
import {ImageMapCoordinate} from "./imagemapcoordinate.model";
export class ProductModel {
  hybrisId: string;
  name: string;
  thumbnailUrl: string;
  formattedPrice: string;
  detailUrl: string;
  description: string;
  coordinates: ImageMapCoordinate;
  interested: boolean
  imageUrl: string

}